hello-spring/src/main/java/Hello/hellospring/service/MemberService.java
package Hello.hellospring.service;

import Hello.hellospring.domain.Member;
import Hello.hellospring.repository.MemberRepository;
import Hello.hellospring.repository.MemoryMemberRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;
//@Service //스프링이 올라올 때 스프링 컨테이너에 등록해준다
public class MemberService { //cmd + sft + T ->Test case 자동 생성
    private final MemberRepository memberRepository;

//    @Autowired
    public MemberService(MemberRepository memberRepository) { //dependency igetion? DI?
        this.memberRepository = memberRepository;
    }
    //회원가입
    public Long join(Member member){
        // 같은 이름이 있는 중복 회원 X
//        Optional<Member> result = memberRepository.findByName(member.getName()); //cmd + option + v
//        result.ifPresent(m -> {
//            throw new IllegalStateException("이미 존재하는 회원입니다.");
//        });
        validateDuplicateMember(member); // ctrl + T ->메서드 추출
        memberRepository.save(member);
        return member.getId();
    }

    private void validateDuplicateMember(Member member) {
        memberRepository.findByName(member.getName())
                .ifPresent(m -> {
                    throw new IllegalStateException("이미 존재하는 회원입니다.");
                });
    }
    //전체 회원 조회
    public List<Member> findMembers(){
        return memberRepository.finaAll();
    }
    public Optional<Member> findOne(Long memberId){
        return memberRepository.findById(memberId);
    }
}
